requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pymongo[srv]==4.7.2
python-dotenv==1.0.1
anthropic==0.25.8
pydantic==2.7.1
